+++ src/CriticalTransitions.jl
@@ -11,7 +11,6 @@ end CriticalTransitions
 using Statistics: Statistics, mean
 using LinearAlgebra: LinearAlgebra, norm, dot, tr, det
 using StaticArrays: StaticArrays, SVector
-using DataStructures: CircularBuffer
 using Random: Random
 
 # Core
@@ -70,7 +69,6 @@ include("largedeviations/methods.jl")
 include("largedeviations/MinimumActionPath.jl")
 include("largedeviations/minimize_action.jl")
 include("largedeviations/minimize_geometric_action.jl")
-
 include("largedeviations/sgMAM.jl")
 include("largedeviations/string_method.jl")